Fixed `yii\web\CacheSession::destroySession()` to work correctly when session is not written yet
Fixes #13537

+++ framework/web/CacheSession.php
@@ -102,6 +102,10 @@ class CacheSession extends Session
      */
     public function destroySession($id)
     {
+        if (!$this->cache->exists($id)) {
+            return true;
+        }
+
         return $this->cache->delete($this->calculateKey($id));
     }
